Move projection tests to a folder
